Check symbolic link exists

+++ install_scripts/openssl.sh
@@ -41,7 +41,9 @@ if [ ! -z $REINSTALL ] || [ -z $INSTALLED_VERSION ] || [ $VER != $INSTALLED_VERS
   curl -L $REPO_URL/archive/$TAG.tar.gz | tar xz && cd $FOLDER
   ./config --prefix=$HOME/.local threads && \
     make -s -j$(nproc) && make -s install_sw 1>/dev/null && make -s install_man_docs 1>/dev/null
-  ln -s $HOME/.local $HOME/.local/.openssl
+  if [ ! -L $HOME/.local/.openssl ];then
+    ln -s $HOME/.local $HOME/.local/.openssl
+  fi
 
   cd $ROOT && rm -rf $TMP_DIR
 else
